api/requirements.txt
flask
flask-cors
flask-limiter
pymysql
python-dotenv
